fix(build): 修复批C后遗留的 8 处编译阻断（机械修复级联损坏）— 全量 native 构建恢复 100%
批C收尾验证时发现 master 存在 8 处 pre-existing 编译阻断（源自 08-07 MISRA 机械修复级联损坏），
全部为 1-2 行最小修复，已通过 build-batchC-final 全量构建验证（BUILD_EXIT=0）：

1. src/bsw/os/include/Std_Types.h: 补 #include <stdint.h>（Dcm.c 的 uint32_t 无法解析，897a72be 引入）
2. config/input/services/E2E_Cfg.h: P2VAR 宏→显式指针类型（119f5edc 移除 P2VAR 后遗留，与 Dem_Lcfg 同款修复）
3. src/telemetry/telemetry.c: (TEL_MAX_EVENT_SIZE -) 括号错位 → 正常减法
4. src/asw/diagnostic_manager/src/Swc_DiagnosticManager.c: &(uint32_t)(request) → &request
5. src/asw/mode_manager/src/Swc_ModeManager.c: 同上
6. tests/mock/test_mcal_uart.c: UART_MAX_CHANNELS 定义移到数组声明前（原为编译错误，测试从未编译通过）
7. src/bsw/boot/src/Boot_Loader.c: whfor (;;) → for (;;) {（d461fd56 级联损坏）
8. src/bsw/services/mqtt/src/Mqtt_Tls.c: (sessionDataSize >=) sizeof() 括号错位 ×2

遗留：mcal_uart_test 编译恢复后暴露 latent SegFault（Init valid），与批C无关，记录在案。

+++ src/asw/diagnostic_manager/src/Swc_DiagnosticManager.c
@@ -523,7 +523,7 @@ void Swc_DiagnosticManager_ProcessRequest(void)
 
     if (!swcDiagManager.hasPendingRequest) {
         /* Check for new request via RTE */
-        if ((uint8_t)(Rte_Read_DiagnosticRequest(&(uint32_t)(request))) == (uint8_t)RTE_E_OK) {
+        if ((uint8_t)(Rte_Read_DiagnosticRequest(&request)) == (uint8_t)RTE_E_OK) {
             swcDiagManager.pendingRequest = request;
             swcDiagManager.hasPendingRequest = TRUE;
         } else {
